+++ src/dune_fmt.mli
@@ -5,3 +5,6 @@ val format_file :
   input:Path.t option ->
   output:Path.t option ->
   unit
+
+(** Pretty-print a list of toplevel s-expressions *)
+val pp_top_sexps : Format.formatter -> Dune_lang.Ast.t list -> unit